requirements.txt
streamlit>=1.28.0
python-dotenv>=1.0.0
tonepilot>=0.1.0 